src/scripts/requirements.txt
requests
pyyaml
beautifulsoup4
html2text